requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
filelock==3.12.2
flake8==6.1.0
flake8-comprehensions==3.14.0
flake8-print==5.0.0
flake8_simplify==0.20.0
frozenlist==1.4.0
gunicorn==21.2.0
identify==2.5.26
idna==3.4
kombu==5.3.1
mccabe==0.7.0
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
phonenumbers==8.13.18
platformdirs==3.10.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.5.0
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
yarl==1.9.2
